test(ui): isolate maintenance payload fixtures
Refs #3019\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
