+++ apps/ui/tests/smoke.update.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
 
-import { createHealthyUpdateStatus } from "./maintenance_feature_test_support";
+import { createHealthyUpdateStatus } from "./maintenance_payload_test_support";
 import {
   bootLiveDashboard,
   fulfillJson,
@@ -367,28 +367,7 @@ test("settings update failure shows retry guidance, failed-stage retention, and 
     });
   });
   await page.route("**/api/health", async (route) => {
-    await fulfillJson(route, {
-      status: "ok",
-      processing_state: "idle",
-      processing_failures: 0,
-      degradation_reasons: [],
-      data_loss: {
-        affected_clients: 0,
-        tracked_clients: 0,
-        frames_dropped: 0,
-        queue_overflow_drops: 0,
-        server_queue_drops: 0,
-        parse_errors: 0,
-      },
-      persistence: {
-        analysis_in_progress: false,
-        analysis_queue_depth: 0,
-        write_error: null,
-        analysis_active_run_id: null,
-        analysis_started_at: null,
-        analysis_elapsed_s: null,
-      },
-    });
+    await fulfillJson(route, createHealthyUpdateStatus());
   });
   await page.route("**/api/update/internet-status", async (route) => {
     await fulfillJson(route, {
